+++ changelog.md
@@ -9,6 +9,7 @@
 	* Flatten and append attribute data when exporting attribute combinations as products
 * Instant search should search manufacturer/brand name
 * (Dev) New 'ApplicationStarted' event
+* **GMC**: Added missing database indexes.
 
 ### Bugfixes
 * #1145: Fixed HTTP 404 after switching language